requirements.txt
pandas==2.3.1
numpy==2.3.2
scipy==1.16.0
torch==2.7.1
gymnasium==1.2.0
stable-baselines3==2.7.0
PyYAML==6.0.2
matplotlib==3.10.3
yfinance==0.2.65
